Remove InteractiveQuestionProviderWithLocation since it's really not necessary
