+++ enterprise/frontend/src/embedding-sdk/components/private/InteractiveQuestion/context/index.ts
@@ -1,3 +1,2 @@
 export * from "./InteractiveQuestionProvider";
-export * from "./InteractiveQuestionProviderWithLocation";
 export * from "./types";
